requirements.txt
# Core dependencies
torch==2.9.1
torchaudio==2.9.1
transformers==4.57.1
einops==0.8.0

# Training
lightning==2.2.1
pytorch-lightning==2.5.0
hydra-core==1.3.2
omegaconf==2.3.0
wandb==0.20.1

# Data processing
datasets==4.2.0
tinytag

# Audio processing
vocos==0.1.0
numpy==1.26.4
librosa==0.11.0
soundfile==0.12.1
torchaudio

# Utilities
click==8.1.7
tqdm==4.67.0
PyYAML==6.0.2

# Performance optimizations
flash-attn==2.8.3
cut-cross-entropy==25.9.3

# Additional dependencies
safetensors==0.4.5
tokenizers==0.22.1
huggingface-hub==0.35.3
pillow==11.0.0
scipy==1.16.3

